don't specify synchronous in enqueue methods
Instead of calling `algolia_synchronous?` and passing the result to the
two enqueue methods, we can just set the default to false. This makes
the methods parallel to the non-enqueue versions of these methods, and
they should ultimately behave the same.

There is a nice side-effect to this change, and that is that it makes
the gem more Rails 2 compatible. The one primary reason it isn't
currently is because the `after_destroy` block calls
`algolia_synchronous?` on `self` implicitly. In Rails 2, these blocks
were called with the model _class_ as the context, whereas in Rails 3+
the model _instance_ is.