Allow configuring user agent / other default options
Currently the default options are merged into the user-supplied
configuration which means that any user supplied configuration for the
user agent would be overwritten.

This could technically be described as a breaking change, but only if
the user was relying on a somewhat odd behaviour. They'd have to set a
configuration value then rely on it being overwritten. I can't think of
a use case for that, but if someone was doing something interesting with
a proxy that sniffs user-agent, and had done the above they might
technically have a breakage. Otherwise, I'd usually describe this as a
bugfix.