Allow library users to specify default ActiveJob queue for indexing
The current auto_indexing behaviours are great, but force a user to
effectively reimplement the default enqueue behaviour if they just wanna
use a different ActiveJob queue. Instead, allow the queue_name to
specified as a configuraiton option, but fallback to the default if no
value is set.

Someone who just updates to a version with this commit should see no
difference until they supply a value for the configuration, if they
wish.