Allow configuration ActiveJob queue for auto_index (#439)
* Specify the GlobalID app name

Required for GlobalID to generate names, but wasn't previously required
as no objects were actually being enqueued and therefore not actually
generating names.

In a rails environment, this would be automatically configured by the
railtie in GlobalID [1].

https://github.com/rails/globalid/blob/b0f5af88e1baf966a4309f97321bf71dcfc6f248/lib/global_id/railtie.rb#L20

* Allow library users to specify default ActiveJob queue for indexing

The current auto_indexing behaviours are great, but force a user to
effectively reimplement the default enqueue behaviour if they just wanna
use a different ActiveJob queue. Instead, allow the queue_name to
specified as a configuraiton option, but fallback to the default if no
value is set.

Someone who just updates to a version with this commit should see no
difference until they supply a value for the configuration, if they
wish.