Fix issue with rules and synonyms loss during full reindex
Previously, when executing a full reindex with a temporary index in the AlgoliaSearch module, rules and synonyms were not copied over to the temporary index, resulting in data loss. This commit addresses the problem by ensuring that rules and synonyms are correctly copied from the source index to the temporary index during a full reindex.